backend/pyproject.toml
[tool.poetry]
name = "en-poisson-backend"
version = "0.1.0"
description = ""
authors = ["nlao1 <nlao@seas.upenn.edu>"]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
fastapi = "^0.109.0"
uvicorn = "^0.26.0"
httpx = "^0.26.0"
black = "^23.12.1"
scrypt = "^0.8.21"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"

[tool.poetry.dev-dependencies]
flake8 = "^7.0.0"
black = "^23.12.1"
isort = "^5.13.2"
flake8-isort = "^6.1.1"
click = "^8.1.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3